canteen_backend/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
sqlparse==0.5.3
